salvando usuário, foto e item no firebase

+++ app/src/main/java/br/com/myfunkos/util/MoedaUtil.java
@@ -1,6 +1,5 @@
 package br.com.myfunkos.util;
 
-import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
 import java.util.Locale;
@@ -12,7 +11,7 @@ public class MoedaUtil {
     public static final String FORMATO_PADRAO = "R$";
     public static final String FORMATO_COM_ESPACO = "R$ ";
 
-    public static String formataParaReal(BigDecimal valor) {
+    public static String formataParaReal(Double valor) {
         NumberFormat formato =
                 DecimalFormat.getCurrencyInstance(new Locale(PORTUGUES, BRASIL));
         return formato
